Introduce protocol_version field for TP registration
SDK_PROTOCOL_VERSION is the current version of SDK.
A internal enum is maintained for the TransactionProcessor
such that when a special feature is requested, the
corresponding protocol version value is sent as
protocol_version field in TpRegisterRequest message.

The request is acknowledged back with a echoed
protocol_version if validator can handle the request
successfully. Otherwise registration attempt would be rejected.

Signed-off-by: S m, Aruna <aruna.s.m@intel.com>
Signed-off-by: S m, Aruna <aruna.mohan@walmartlabs.com>