Allow additional protocols for consensus engines
Adds an (optional) list of additional protocols that a consensus engine
supports as part of the registration process. These additional protocols
can be used for backwards compatiblity. If the values of the
`sawtooth.consensus.algorithm.name` and
`sawtooth.consensus.algorithm.version` settings match the name/version
of an engine or the name/version pair is in the engine's list of
additional protocols, that engine will be activated.

Signed-off-by: Logan Seeley <seeley@bitwise.io>