Fix clippy::single_range_in_vec_init lint
This is almost always incorrect, as it will result in a
Vec that has only one element. Almost always, the programmer
intended for it to include all elements in the range or
for the end of the range to be the length instead.

This means our correlations ids are not 16 characters long
like we want. This fixes that issue.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>